training/requirements.txt
torch>=2.2
transformers>=4.39
datasets>=2.18
accelerate>=0.28
evaluate>=0.4
jiwer>=3.0
opencv-python>=4.9
Pillow>=10.3
